Update to ssh library and bug fixes.
* Replaced the ssh library with [sshj](https://github.com/hierynomus/sshj)
    * Adds support for new ssh crypto algorithms
    * Adds support for FIPS compliant algorithms
    * More modern ssh implementation
* Added support for importing and building the library with maven
* Added FindBugs code testing to maven build

This is a major change to the netconf-java library. This version changes how Devices are
created. Instead of passing in parameters in a constructor, Device now uses a Builder. This
allows passing in whichver parameters you need and creating a unique Device object. Once created,
you no longer change change attributes like userName or port. You would need to create a new Device object.

See the Examples/CreateDevice class for examples on how to create a new Device.